requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --output-file requirements.txt
aiofiles==25.1.0
    # via aiogram
aiogram==3.23.0
    # via
    #   aiogram-dialog
    #   todo-bot
aiogram-dialog==2.4.0
    # via todo-bot
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   aiogram
    #   todo-bot
aiosignal==1.4.0
    # via aiohttp
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.0
    # via django
attrs==25.4.0
    # via aiohttp
billiard==4.2.4
    # via celery
cachetools==5.5.2
    # via
    #   aiogram-dialog
    #   google-auth
celery==5.6.0
    # via
    #   django-celery-beat
    #   todo-bot
certifi==2025.11.12
    # via
    #   aiogram
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   timezonefinder
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
cron-descriptor==2.0.6
    # via django-celery-beat
cryptography==46.0.3
    # via
    #   types-pyopenssl
    #   types-redis
dateparser==1.2.2
    # via todo-bot
django==5.2.9
    # via
    #   django-celery-beat
    #   django-stubs
    #   django-stubs-ext
    #   django-timezone-field
    #   djangorestframework
    #   todo-bot
django-celery-beat==2.8.1
    # via todo-bot
django-stubs==5.2.8
    # via djangorestframework-stubs
django-stubs-ext==5.2.8
    # via django-stubs
django-timezone-field==7.2.1
    # via django-celery-beat
djangorestframework==3.16.1
    # via todo-bot
djangorestframework-stubs==3.16.6
exceptiongroup==1.3.1
    # via celery
flatbuffers==25.9.23
    # via timezonefinder
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2 ; python_full_version >= '3.14'
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-core==2.28.1 ; python_full_version < '3.14'
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.45.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via todo-bot
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==23.0.0
    # via todo-bot
h11==0.16.0
    # via uvicorn
h3==4.4.1
    # via timezonefinder
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via aiogram-dialog
kombu==5.6.1
    # via celery
librt==0.7.4 ; platform_python_implementation != 'PyPy'
    # via mypy
magic-filter==1.0.12
    # via aiogram
markupsafe==3.0.3
    # via jinja2
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy==1.19.1
mypy-extensions==1.1.0
    # via mypy
numpy==2.3.5
    # via timezonefinder
packaging==25.0
    # via
    #   gunicorn
    #   kombu
pathspec==0.12.1
    # via mypy
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via todo-bot
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   aiogram
    #   google-generativeai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.2.5
    # via httplib2
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   dateparser
    #   todo-bot
python-decouple==3.8
    # via todo-bot
pytz==2025.2
    # via
    #   dateparser
    #   todo-bot
redis==7.1.0
    # via todo-bot
regex==2025.11.3
    # via dateparser
requests==2.32.5
    # via
    #   djangorestframework-stubs
    #   google-api-core
    #   todo-bot
rsa==4.9.1
    # via google-auth
ruff==0.14.9
six==1.17.0
    # via python-dateutil
sqlparse==0.5.4
    # via django
timezonefinder==8.1.0
    # via todo-bot
tqdm==4.67.1
    # via google-generativeai
types-cffi==1.17.0.20250915
    # via types-pyopenssl
types-pyopenssl==24.1.0.20240722
    # via types-redis
types-pyyaml==6.0.12.20250915
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-redis==4.6.0.20241004
types-requests==2.32.4.20250913
    # via djangorestframework-stubs
types-setuptools==80.9.0.20250822
    # via types-cffi
typing-extensions==4.15.0
    # via
    #   aiogram
    #   aiosignal
    #   cron-descriptor
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   exceptiongroup
    #   google-generativeai
    #   grpcio
    #   mypy
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via
    #   django
    #   django-celery-beat
    #   kombu
    #   tzlocal
tzlocal==5.3.1
    # via
    #   celery
    #   dateparser
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via
    #   requests
    #   types-requests
uvicorn==0.38.0
    # via todo-bot
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
yarl==1.22.0
    # via aiohttp
